fix problems introduced by using url.resolve in HAL walker
Before this commit, post processing the url happened before resolving URI
templates. In node 0.10 and browserify with url@0.7.9 this is okay, but in the
near future (node 0.11/0.12 and url@0.10.1) this will break because curly braces
will be url encoded and then the URI template resolving will not work.

Therefore, we swap the order and resolve the URI template before post processing
the URI.