This gracefully handles requests with now 'follows' in them
The issue here is that a 'normal' request will follow relations.  But there are cases
(e.g., getting information about the root URI or creating a request based on a Location
header) where no relation needs to be followed.  Currently, traverson requires an
empty 'follow()' invocation or it fails.  This simple fix eliminates the requirement
for this empty call.